Use non-blocking queue calls for add and blocking calls for get
The idea behind this is that the inference module always wants to wait
for the next element in the queue since that will be the most recent
analysis parameter. The OSC listener, however, just wants to put its
newly updated values to the queue as fast as possible so that the
inference module can make use of it. For this reason, it also makes
sense to use a max queue size of 1, since we never want to "go back in
time" to retreive old values from the queue.